+++ src/test/java/blackjack/card/CardDeckTest.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.card;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardDeck;
